Exercise 50 complete.

+++ Exercise50/Triangle.cs
@@ -6,32 +6,32 @@ namespace Exercise50
 {
     public class Triangle
     {
-        private int sideLength1 { get; set;}
-        public int SideLength1
+        private double sideLength1 { get; set;}
+        public double SideLength1
         {
             get { return sideLength1; }
             set { sideLength1 = SideLength1; }
         }
-        private int sideLength2 { get; set; }
-        public int SideLength2
+        private double sideLength2 { get; set; }
+        public double SideLength2
         {
             get { return sideLength2; }
             set { sideLength2 = SideLength2; }
         }
-        private int sideLength3 { get; set; }
-        public int SideLength3
+        private double sideLength3 { get; set; }
+        public double SideLength3
         {
             get { return sideLength3; }
             set { sideLength3 = SideLength3; }
         }
 
-        public double AreaOfTriangle (int sideLength1, int sideLength2)
+        public double AreaOfTriangle (double sideLength1, double sideLength2)
         {
             double areaOfTriangle = (sideLength1 + sideLength2) / 2;
             return areaOfTriangle;
         }
 
-        public double PerimeterOfTriangle (int sideLength1, int sideLength2, int sideLength3)
+        public double PerimeterOfTriangle (double sideLength1, double sideLength2, double sideLength3)
         {
             double perimeterOfTriangle = sideLength1 + sideLength2 + sideLength3;
             return perimeterOfTriangle;
